Add files via upload
The VisionEcho project is a smart system that helps blind and visually impaired people understand what is shown in an image. It does this in two main steps: first, it looks at a picture and creates a sentence describing what’s happening (like “a man is riding a bike on the street”). Then, it reads that sentence out loud using a voice tool. This means someone who can’t see the image can still know what’s in it just by listening.

We use a special deep learning model to do this. It includes a CNN (InceptionV3) to see and understand the image, and an RNN (GRU) to write a caption about the image. We also use an attention mechanism so the model knows which part of the image to focus on when writing each word. Finally, we use a Text-to-Speech (TTS) tool to speak the caption.

This project was built using Python with TensorFlow and other helpful libraries like NumPy, Pandas, and gTTS. The data used for training came from the Flickr8k dataset, which includes thousands of images and their descriptions. The final goal of this project is to make visual information more accessible for people who cannot see.